Creando componentes en agregar y listado de pagos
