--- src/components/layout/backupProfile.vue
@@ -1,160 +1,0 @@
-<template lang="pug">
-  main
-    .container
-      .row.profile
-        .col-md-3
-          .profile-sidebar
-            .profile-userpic
-              img.img-responsive(:src='clientes.ruta_imagen')
-            .profile-usertitle
-              .profile-usertitle-name {{ clientes.nombre }}
-              .profile-usertitle-job {{ clientes.telefono }}
-            .profile-userbuttons
-              button.btn.btn-success.btn-sm(@click='hide') Cerrar
-              button.btn.btn-danger.btn-sm(@click='test') Message
-            .profile-usermenu
-              ul.nav
-                li.active
-                  a(href='#')
-                    i.glyphicon.glyphicon-home
-                    |  {{ clientes.nombre }}
-                li
-                  a(href='#')
-                    i.glyphicon.glyphicon-user
-                    | {{ clientes.telefono }}
-                li
-                  a(href='#', target='_blank')
-                    i.glyphicon.glyphicon-ok
-                    | {{ clientes.correo }}
-                li
-                  a(href='#')
-                    i.glyphicon.glyphicon-flag
-                    | {{ clientes.mensualidad }}
-</template>
-
-<script>
-export default {
-  computed: {
-    Client () {
-      this.cliente = this.$store.state.clientSelect.nombre
-    }
-  },
-  props: {
-    clientes: { type: Object, required: true }
-  },
-  methods: {
-    hide () {
-      this.$store.commit('hide')
-    },
-    test () {
-      console.log(this.$store.state.clientSelect)
-    }
-  }
-}
-</script>
-
-<style lang="scss" scoped>
-.profile {
-  margin: 20px 0;
-}
-
-/* Profile sidebar */
-.profile-sidebar {
-  padding: 20px 0 10px 0;
-  background: #fff;
-}
-
-.profile-userpic img {
-  float: none;
-  margin: 0 auto;
-  width: 50%;
-  height: 50%;
-  -webkit-border-radius: 50% !important;
-  -moz-border-radius: 50% !important;
-  border-radius: 50% !important;
-}
-
-.profile-usertitle {
-  text-align: center;
-  margin-top: 20px;
-}
-
-.profile-usertitle-name {
-  color: #5a7391;
-  font-size: 16px;
-  font-weight: 600;
-  margin-bottom: 7px;
-}
-
-.profile-usertitle-job {
-  text-transform: uppercase;
-  color: #5b9bd1;
-  font-size: 12px;
-  font-weight: 600;
-  margin-bottom: 15px;
-}
-
-.profile-userbuttons {
-  text-align: center;
-  margin-top: 10px;
-}
-
-.profile-userbuttons .btn {
-  text-transform: uppercase;
-  font-size: 11px;
-  font-weight: 600;
-  padding: 6px 15px;
-  margin-right: 5px;
-}
-
-.profile-userbuttons .btn:last-child {
-  margin-right: 0px;
-}
-    
-.profile-usermenu {
-  margin-top: 30px;
-}
-
-.profile-usermenu ul li {
-  border-bottom: 1px solid #f0f4f7;
-}
-
-.profile-usermenu ul li:last-child {
-  border-bottom: none;
-}
-
-.profile-usermenu ul li a {
-  color: #93a3b5;
-  font-size: 14px;
-  font-weight: 400;
-}
-
-.profile-usermenu ul li a i {
-  margin-right: 8px;
-  font-size: 14px;
-}
-
-.profile-usermenu ul li a:hover {
-  background-color: #fafcfd;
-  color: #5b9bd1;
-}
-
-.profile-usermenu ul li.active {
-  border-bottom: none;
-}
-
-.profile-usermenu ul li.active a {
-  color: #5b9bd1;
-  background-color: #f6f9fb;
-  border-left: 2px solid #5b9bd1;
-  margin-left: -2px;
-}
-
-/* Profile Content */
-.profile-content {
-  padding: 20px;
-  background: #fff;
-  min-height: 460px;
-}
-
-</style>
